fix: fix bug with backdrop and slider
