+++ src/components/product-paper/index.ts
@@ -1,0 +1,1 @@
+export * from './product-paper';
